refactor: move server/ws_approval_gate.rs to astra-turn-core (232 LOC)
Move the WebSocket-based tool approval gate to astra-turn-core, where the
edge_ledger dependency already lives.

Changes:
- Copy ws_approval_gate.rs to astra-turn-core (232 LOC)
- Add astra-tools dependency to astra-turn-core
- Register module in lib.rs
- Replace original with re-export stub
