+++ rust/crates/astra-turn-core/src/lib.rs
@@ -166,3 +166,4 @@ pub mod result_quality {
     pub use astra_turn_types::{ResultQuality, classify_result, quality_feedback};
 }
 pub mod turn_guard;
+pub mod ws_approval_gate;
